Properly link against mathic and memtailor.
In 691edb2, the use of pkg-config to get the library names of mathic and
memtailor was dropped.  Instead, they were manually added when linking
the mgb binary and the unit tests.

However, they were not manually added to the linking of the libmathicgb
library, causing underlinking.  For example, from the build logs of the
Debian package:

    dpkg-shlibdeps: warning: symbol _ZN6mathic13ColumnPrinter11bytesInUnitB5cxx11Ey used by debian/libmathicgb0/usr/lib/i386-linux-gnu/libmathicgb.so.0.0.0 found in none of the libraries

In this patch, we instead define MATHIC_LIBS and MEMTAILOR_LIBS, which
used to be defined by pkg-config.  The library names are then included
with DEPS_LIBS.