Check for googletest version >= 1.10.0
We try compiling a small program that uses TYPED_TEST_SUITE, which was
introduced in this version.  If it doesn't compile, then we must have
a lower version, and so we download googletest instead.

We also take the opportunity to print a few more useful messages,
quiet the output of unzip, and check that we have write permission
in the directory we're trying to download googletest into.