Call the destructor explicitly when using placment new
Lookup has a non-trivial destructor, so the failure to call it's destructor in the destructor of ConcreteModuleMonoSet was causing memory leaks. I added some other calls to destructors in other places that use placement new, but not all of them. There are a couple places where the object involved is trivially destructable, and calling the destructors in the right places is not trivial.